Modified model to show cyrillic characters. Deleted uselles class

+++ WeaponAccountingProject/ViewModels/CreateWeaponViewModel.cs
@@ -7,7 +7,7 @@ namespace WeaponAccountingProject.ViewModels
         public string Name { get; set; } = null!;
         public string Year { get; set; } = null!;
         public string? RecordNumber { get; set; } = null!;
-        public int Value { get; set; } = 0;
+        public decimal Value { get; set; } = 0;
         public int LocationId { get; set; } 
     }
 }
