+++ WeaponAccountingProject/ViewModels/EditWeaponViewModel.cs
@@ -6,7 +6,7 @@
         public string Name { get; set; } = null!;
         public string Year { get; set; } = null!;
         public string? RecordNumber { get; set; } = null!;
-        public int Value { get; set; } = 0;
+        public decimal Value { get; set; } = 0;
         public int? LocationId { get; set; }
     }
 }
